ADS: Variables: Handle struct writing
Actually, this probably won't be particularly useful because an object
with data type `Variable::Struct` may not necessarily be a `STRUCT`; for
example, it could be a `FUNCTION_BLOCK`. Interspersed within the data
type may be bytes of memory which are used by the system but not exposed
via ADS; for example, a `METHOD` temporary variable. It could be a very
bad idea to try to write to those bytes! Therefore, if the input data
does not translate to contiguous bytes, the function will return an
error.

Anyway, it is easy for the caller to iterate through the struct fields
and write each field member individually.